v13.3: fix BF16/FP32 dtype mismatch in CfC layer's scatter-add
Under BF16 autocast:
  msg     = self.msg_mlp(...)                      # BF16 (Linear autocast)
  msg     = msg * edge_weight.unsqueeze(0)...      # BF16 * FP32 → FP32 (promotion)
  agg     = torch.zeros(..., dtype=h.dtype)        # BF16
  agg.index_add(1, dst, msg)                       # ERROR: BF16 self, FP32 source

Fix: cast msg back to h.dtype after the multiply so it matches agg.  Edge
weights stay FP32 (right for storage; we add gibbs weighting later) but
get matched on the multiply.  Reproduced with a direct unit-test of
_CfCGraphLayer._forward_impl with BF16 input + FP32 edge_weight; now
returns BF16 cleanly.

Doesn't affect the integration test (CPU path, no autocast).

+++ colab_cell_emulator.py
@@ -1409,7 +1409,10 @@ class _CfCGraphLayer(nn.Module):
         h_src = h.index_select(1, src)        # (B, E, H)
         h_dst = h.index_select(1, dst)
         msg = self.msg_mlp(torch.cat([h_src, h_dst], dim=-1))
-        msg = msg * edge_weight.unsqueeze(0).unsqueeze(-1)
+        # v13.3: cast back to h.dtype.  Under BF16 autocast msg is BF16, but
+        # edge_weight is an fp32 buffer, so `msg * edge_weight` gets promoted
+        # to fp32 — the index_add into a BF16 agg below then errors.
+        msg = (msg * edge_weight.unsqueeze(0).unsqueeze(-1)).to(h.dtype)
         agg = torch.zeros(B, N, H, device=h.device, dtype=h.dtype)
         agg = agg.index_add(1, dst, msg)      # out-of-place: autograd-safe
         # Degree-normalise so high-degree nodes don't blow up
